Added tests for invalid type and for erring in dev mode for soap flag
Also some refactoring
